web/requirements.txt
django>=4.2,<5.0
gunicorn
boto3
django-dotenv
psycopg2-binary
django-storages
requests
django-extensions==3.2.3
django-ckeditor==6.7.0
django-js-asset==2.1.0